refactor: constants 폴더의 구조를 변경한다.
