+++ src/constants/chargeMoney/action.ts
@@ -1,0 +1,3 @@
+const CHARGE_VENDING_MONEY = 'CHARGE_MONEY'
+
+export { CHARGE_VENDING_MONEY }
