+++ src/constants/chargeMoney/element.ts
@@ -1,0 +1,11 @@
+const VENDING_MACHINE_CHARGE_INPUT = '#vending-machine-charge-input'
+const VENDING_MACHINE_CHARGE_BUTTON = '#vending-machine-charge-button'
+const VENDING_MACHINE_COIN_INVENTORY = '#vending-coin-inventory'
+const VENDING_MACHINE_CHARGE_AMOUNT = '#vending-machine-charge-amount'
+
+export {
+  VENDING_MACHINE_CHARGE_INPUT,
+  VENDING_MACHINE_CHARGE_BUTTON,
+  VENDING_MACHINE_COIN_INVENTORY,
+  VENDING_MACHINE_CHARGE_AMOUNT,
+}
