+++ src/constants/chargeMoney/errorMessage.ts
@@ -1,0 +1,10 @@
+const VENDING_MACHINE_CHARGE_COIN_INPUT_EMPTY = '금액을 입력해주세요.'
+const VENDING_MACHINE_CHARGE_COIN_MINIMUM_INPUT_INVALID =
+  '금액을 100원 이상 입력해주세요.'
+const VENDING_MACHINE_CHARGE_COIN_INPUT_SPLIT_INVALID =
+  '금액을 10원 단위로 나누어 떨어져야 합니다.'
+export {
+  VENDING_MACHINE_CHARGE_COIN_INPUT_EMPTY,
+  VENDING_MACHINE_CHARGE_COIN_MINIMUM_INPUT_INVALID,
+  VENDING_MACHINE_CHARGE_COIN_INPUT_SPLIT_INVALID,
+}
