+++ src/constants/productManage/action.ts
@@ -1,0 +1,3 @@
+const ADD_PRODUCT = 'ADD_PRODUCT'
+
+export { ADD_PRODUCT }
