+++ src/constants/productManage/element.ts
@@ -1,0 +1,17 @@
+const PRODUCT_NAME_INPUT = '#product-name-input'
+const PRODUCT_PRICE_INPUT = '#product-price-input'
+const PRODUCT_QUANTITY_INPUT = '#product-quantity-input'
+const PRODUCT_INVENTORY_CONTAINER = '#product-inventory-container'
+const PRODUCT_ADD_BUTTON = '#product-add-button'
+const PRODUCT_CONTAINER = '#product-container'
+const PRODUCT_INVENTORY = '#product-inventory'
+
+export {
+  PRODUCT_NAME_INPUT,
+  PRODUCT_PRICE_INPUT,
+  PRODUCT_QUANTITY_INPUT,
+  PRODUCT_CONTAINER,
+  PRODUCT_INVENTORY,
+  PRODUCT_ADD_BUTTON,
+  PRODUCT_INVENTORY_CONTAINER,
+}
