+++ src/constants/productManage/errorMessage.ts
@@ -1,0 +1,10 @@
+const PRODUCT_ADD_INPUT_INVALID = '상품 입력창에 값을 압력해주세요.'
+const PRODUCT_ADD_PRICE_INVALID =
+  '상품의 금액은 100원 이상이어야 하고 10원으로 나누어 떨어져야 합니다.'
+const PRODUCT_ADD_QUANTITY_INVALID = '상품의 수량은 1개 이상이어야 합니다.'
+
+export {
+  PRODUCT_ADD_INPUT_INVALID,
+  PRODUCT_ADD_PRICE_INVALID,
+  PRODUCT_ADD_QUANTITY_INVALID,
+}
