+++ src/constants/productPurchase/action.ts
@@ -1,0 +1,5 @@
+const ADD_USER_MONEY = 'ADD_USER_MONEY'
+const PURCHASE_PRODUCT = 'PURCHASE_PRODUCT'
+const RETURN_COIN = 'RETURN_COIN'
+
+export { ADD_USER_MONEY, PURCHASE_PRODUCT, RETURN_COIN }
