+++ src/constants/productPurchase/element.ts
@@ -1,0 +1,15 @@
+const PRODUCT_PURCHASE_RETURN_COIN_INVENTORY = '#return-coin-inventory'
+const PRODUCT_PURCHASE_PRODUCT_INPUT_CONTAINER = '#product-items-container'
+const PRODUCT_PURCHASE_CHARGE_INPUT = '#charge-input'
+const PRODUCT_PURCHASE_CHARGE_BUTTON = '#charge-button'
+const PRODUCT_PURCHASE_CHARGE_AMOUNT = '#charge-amount'
+const PRODUCT_PURCHASE_COIN_RETURN_BUTTON = '#coin-return-button'
+
+export {
+  PRODUCT_PURCHASE_RETURN_COIN_INVENTORY,
+  PRODUCT_PURCHASE_PRODUCT_INPUT_CONTAINER,
+  PRODUCT_PURCHASE_CHARGE_INPUT,
+  PRODUCT_PURCHASE_CHARGE_BUTTON,
+  PRODUCT_PURCHASE_CHARGE_AMOUNT,
+  PRODUCT_PURCHASE_COIN_RETURN_BUTTON,
+}
