+++ src/constants/productPurchase/errorMessage.ts
@@ -1,0 +1,17 @@
+const PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_EMPTY = '충전 할 금액을 입력해주세요.'
+const PURCHASE_PRODUCT_CHARGE_MONEY_MINIMUM_INPUT_INVALID =
+  '10원 이상의 금액을 충전해주세요.'
+const PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_SPLIT_INVALID =
+  '충전 금액은 10원 단위로 나누어 떨어져야 합니다.'
+
+const PURCHASE_PRODUCT_PURCHASE_NO_MONEY =
+  '상품을 구매하기 위한 금액이 부족합니다.'
+
+const PURCHASE_PRODUCT_PURCHASE_PRODUCT_SOLD_OUT = '상품이 매진되었습니다.'
+export {
+  PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_EMPTY,
+  PURCHASE_PRODUCT_CHARGE_MONEY_MINIMUM_INPUT_INVALID,
+  PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_SPLIT_INVALID,
+  PURCHASE_PRODUCT_PURCHASE_PRODUCT_SOLD_OUT,
+  PURCHASE_PRODUCT_PURCHASE_NO_MONEY,
+}
