+++ src/controller/chargeMoney/chargeMoneyController.ts
@@ -1,6 +1,6 @@
+import { CHARGE_VENDING_MONEY } from '../../constants/chargeMoney/action'
 import MoneyChargeModel from '../../model/moneyChargeModel'
 import MoneyStore from '../../store/MoneyStore'
-import { CHARGE_VENDING_MONEY } from '../../utils/constants/action'
 import { ChargeMoneyReucerAction } from './chargeMoneyAction'
 
 export default class ChargeMoneyController {
