--- src/utils/constants/action.ts
@@ -1,13 +1,0 @@
-const ADD_PRODUCT = 'ADD_PRODUCT'
-const CHARGE_VENDING_MONEY = 'CHARGE_MONEY'
-const ADD_USER_MONEY = 'ADD_USER_MONEY'
-const PURCHASE_PRODUCT = 'PURCHASE_PRODUCT'
-const RETURN_COIN = 'RETURN_COIN'
-
-export {
-  ADD_PRODUCT,
-  CHARGE_VENDING_MONEY,
-  ADD_USER_MONEY,
-  PURCHASE_PRODUCT,
-  RETURN_COIN,
-}
