--- src/utils/constants/element.ts
@@ -1,41 +1,0 @@
-// 상품 관리
-const PRODUCT_NAME_INPUT = '#product-name-input'
-const PRODUCT_PRICE_INPUT = '#product-price-input'
-const PRODUCT_QUANTITY_INPUT = '#product-quantity-input'
-const PRODUCT_INVENTORY_CONTAINER = '#product-inventory-container'
-const PRODUCT_ADD_BUTTON = '#product-add-button'
-const PRODUCT_CONTAINER = '#product-container'
-const PRODUCT_INVENTORY = '#product-inventory'
-
-// 잔돈 충전
-const VENDING_MACHINE_CHARGE_INPUT = '#vending-machine-charge-input'
-const VENDING_MACHINE_CHARGE_BUTTON = '#vending-machine-charge-button'
-const VENDING_MACHINE_COIN_INVENTORY = '#vending-coin-inventory'
-const VENDING_MACHINE_CHARGE_AMOUNT = '#vending-machine-charge-amount'
-
-const PRODUCT_PURCHASE_RETURN_COIN_INVENTORY = '#return-coin-inventory'
-const PRODUCT_PURCHASE_PRODUCT_INPUT_CONTAINER = '#product-items-container'
-const PRODUCT_PURCHASE_CHARGE_INPUT = '#charge-input'
-const PRODUCT_PURCHASE_CHARGE_BUTTON = '#charge-button'
-const PRODUCT_PURCHASE_CHARGE_AMOUNT = '#charge-amount'
-const PRODUCT_PURCHASE_COIN_RETURN_BUTTON = '#coin-return-button'
-
-export {
-  PRODUCT_NAME_INPUT,
-  PRODUCT_PRICE_INPUT,
-  PRODUCT_QUANTITY_INPUT,
-  PRODUCT_CONTAINER,
-  PRODUCT_INVENTORY,
-  PRODUCT_ADD_BUTTON,
-  PRODUCT_INVENTORY_CONTAINER,
-  VENDING_MACHINE_CHARGE_INPUT,
-  VENDING_MACHINE_CHARGE_BUTTON,
-  VENDING_MACHINE_COIN_INVENTORY,
-  VENDING_MACHINE_CHARGE_AMOUNT,
-  PRODUCT_PURCHASE_RETURN_COIN_INVENTORY,
-  PRODUCT_PURCHASE_PRODUCT_INPUT_CONTAINER,
-  PRODUCT_PURCHASE_CHARGE_INPUT,
-  PRODUCT_PURCHASE_CHARGE_BUTTON,
-  PRODUCT_PURCHASE_CHARGE_AMOUNT,
-  PRODUCT_PURCHASE_COIN_RETURN_BUTTON,
-}
