--- src/utils/constants/errorMessage.ts
@@ -1,35 +1,0 @@
-// product manage error message
-const PRODUCT_ADD_INPUT_INVALID = '상품 입력창에 값을 압력해주세요.'
-const PRODUCT_ADD_PRICE_INVALID =
-  '상품의 금액은 100원 이상이어야 하고 10원으로 나누어 떨어져야 합니다.'
-const PRODUCT_ADD_QUANTITY_INVALID = '상품의 수량은 1개 이상이어야 합니다.'
-
-const VENDING_MACHINE_CHARGE_COIN_INPUT_EMPTY = '금액을 입력해주세요.'
-const VENDING_MACHINE_CHARGE_COIN_MINIMUM_INPUT_INVALID =
-  '금액을 100원 이상 입력해주세요.'
-const VENDING_MACHINE_CHARGE_COIN_INPUT_SPLIT_INVALID =
-  '금액을 10원 단위로 나누어 떨어져야 합니다.'
-
-const PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_EMPTY = '충전 할 금액을 입력해주세요.'
-const PURCHASE_PRODUCT_CHARGE_MONEY_MINIMUM_INPUT_INVALID =
-  '10원 이상의 금액을 충전해주세요.'
-const PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_SPLIT_INVALID =
-  '충전 금액은 10원 단위로 나누어 떨어져야 합니다.'
-
-const PURCHASE_PRODUCT_PURCHASE_NO_MONEY =
-  '상품을 구매하기 위한 금액이 부족합니다.'
-
-const PURCHASE_PRODUCT_PURCHASE_PRODUCT_SOLD_OUT = '상품이 매진되었습니다.'
-export {
-  PRODUCT_ADD_INPUT_INVALID,
-  PRODUCT_ADD_PRICE_INVALID,
-  PRODUCT_ADD_QUANTITY_INVALID,
-  VENDING_MACHINE_CHARGE_COIN_INPUT_EMPTY,
-  VENDING_MACHINE_CHARGE_COIN_MINIMUM_INPUT_INVALID,
-  VENDING_MACHINE_CHARGE_COIN_INPUT_SPLIT_INVALID,
-  PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_EMPTY,
-  PURCHASE_PRODUCT_CHARGE_MONEY_MINIMUM_INPUT_INVALID,
-  PURCHASE_PRODUCT_CHARGE_MONEY_INPUT_SPLIT_INVALID,
-  PURCHASE_PRODUCT_PURCHASE_NO_MONEY,
-  PURCHASE_PRODUCT_PURCHASE_PRODUCT_SOLD_OUT,
-}
